refactor : GameRule 상수클래스 이름 변경
