+++ src/main/java/bridge/util/GameRule.java
@@ -1,6 +1,6 @@
 package bridge.util;
 
-public class BridgeRule {
+public class GameRule {
 
     public static final int BRIDGE_LENGTH_MIN = 3;
     public static final int BRIDGE_LENGTH_MAX = 20;
